Made safer but slightly slower

+++ pegbis/segment.py
@@ -6,4 +6,4 @@ except ImportError:
 
 
 def segment_graph(edge_index, edge_score, num_vertices, c, min_size=-1):
-    return sg(edge_index, torch.argsort(edge_score), edge_score, num_vertices, c, min_size)
+    return sg(edge_index, torch.argsort(edge_score).to(torch.int32), edge_score.to(torch.float32), num_vertices, c, min_size)
